requirements.txt
nose>=1.3.7
PyHamcrest>=1.9.0
PyGObject>=2.10
schedule>=0.3.2
mockito>=0.7.0
